Remove unused models_improved.py and enhance models.py imports
The `models_improved.py` file was fully removed as it was deemed unnecessary, while `models.py` was updated with the relevant functionality and import improvements. This consolidation simplifies the codebase and eliminates redundancy.
